NavBar and Hero Complete
I have completed the landing page header. Additionally I completed the form that goes on top of the hero image. All of it is mobile compatible to a degree, in order to be truly mobile compatible a media query would be necessary. Also made the date sections with a dropdown so the user can click a calendar!